refactor: move json file functions to json file manager
